requirements.txt
baidu-aip==1.6.4.0
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
olefile==0.44
Pillow==4.2.1
requests==2.18.4
urllib3==1.22
